Refactor: API Response 변경
