+++ shop/src/main/java/max/shop/dto/response/MultipleResult.java
@@ -1,4 +1,4 @@
-package max.shop.response;
+package max.shop.dto.response;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -8,5 +8,5 @@ import java.util.List;
 @Getter
 @Setter
 public class MultipleResult<T> extends Result{
-    List<T> data;
+    List<T> datas;
 }
