--- shop/src/main/java/max/shop/response/Result.java
@@ -1,12 +1,0 @@
-package max.shop.response;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class Result {
-    private boolean success = false;
-    private int code = 200;
-    private String message;
-}
